requirements.txt
PyQt5~=5.15.4
opencv-python~=4.5.2.52
numpy~=1.20.3

Pillow~=8.2.0
tqdm~=4.61.0
torch~=1.8.1
torchsummary~=1.5.1
torchvision~=0.9.1+cu111